refactor: move search params to other fn
