+++ src/search.js
@@ -39,3 +39,43 @@ export function getFixedFilters(
 
     return params;
 }
+
+export function generateSearchParams(
+    searchOptions,
+    searchableFields,
+    dateOptions
+) {
+    let params = "";
+    const fixedFilters = Object.keys(searchOptions.fixedFilters || {});
+    const dateFilters = Object.keys(searchOptions.dates || {});
+    let searchFields = [];
+    searchOptions.text = searchOptions.text.trim();
+
+    if (searchOptions.text.length) {
+        if (!searchOptions.filters.length) {
+            searchFields = searchableFields.filter(field => {
+                return (
+                    !fixedFilters.includes(field) &&
+                    !dateFilters.includes(field)
+                );
+            });
+        } else {
+            searchFields = searchOptions.filters.filter(field => {
+                return (
+                    !fixedFilters.includes(field) &&
+                    !dateFilters.includes(field)
+                );
+            });
+        }
+        params += getParams(searchFields, searchOptions);
+    }
+
+    params = getFixedFilters(searchOptions, params, {
+        formatDate: dateOptions.formatDate,
+        mainDateField: dateOptions.mainDateField
+    });
+
+    return {
+        q: `(${params})`
+    };
+}
